+++ lib/compony/request_context.rb
@@ -46,8 +46,14 @@ module Compony
     def content(name)
       name = name.to_sym
       content_block = component.content_blocks.find { |el| el.name == name } || fail("Content block #{name.inspect} not found in #{component.inspect}.")
-      # A fresh RequestContext is needed due to Rails' buffer
-      concat Compony::RequestContext.new(component, controller, helpers:, locals: local_assigns).evaluate(&content_block.payload)
+      # We have to clear Rails' output_buffer to prevent double rendering of blocks. To achieve this, a fresh render context is instanciated.
+      concat controller.render_to_string(
+        type:   :dyny,
+        locals: { render_component: component, render_controller: controller, render_locals: local_assigns, render_block: content_block },
+        inline: <<~RUBY
+          Compony::RequestContext.new(render_component, render_controller, helpers: self, locals: local_assigns).evaluate_with_backfire(&render_block.payload)
+        RUBY
+      )
     end
   end
 end